backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pydantic
pydantic-settings
google-genai
requests